Restructure clip variant into multi-zoom visualization notebook
Rename the clip2000 trellis to -zoom and refactor it into a CLI-arg-free
visualization notebook that renders three zoom windows from a shared
make_trellis helper: the first 2000, 400, and 200 updates. Hard-code the
OSF slugs (no mo.cli_args) and stamp the clip window into the teeplot
output attrs so the three figures land in distinct files.
